bench  2236104

+++ src/board.h
@@ -43,7 +43,7 @@ public:
     uint64_t threat();
 
     uint64_t key;
-    uint16_t nnueKey;
+    uint64_t nnueKey;
     uint64_t pawnKey;
     uint64_t nonPawnKey[2];
     uint64_t bitboards[N_PIECES];
